Preview/prometheus metrics (#13)
* Arbitrum on deployment configuration

* Scripts to add new token to the bridge

* Stopped tracking sepolia.json

* ignoring config files

* Fixing one cent bug
Sometimes the convertion of a token with more decimals than hathor decimals is going to generate a extra cent because using Math.ceil. Now we just decided to ignore the extra decimals, so anything beyond the 2 hathor decimals is going to be ignored. Also, adding a nice script to deploy multiple federators locally.

* Fixing unhandled exception breaking federator

* Adding prometheus metric push to federator

* Fixing /metrics endpoint return type

* Fixing bug not ending write stream

* Fixing type of response on /metrics

* Changing metric definition

* Adding observability